Added to javascript file:
Today I added an eraser, pencil and cross hairs cursor, I added code to erase the cells the user can click on and I added code to make the cell white when you click a cell while pressing alt. Lastly we made it so we can see the solution of the puzzle.